Add Staking Component
