+++ StratisCore.UI/src/app/wallet/dashboard/dashboard.component.html
@@ -5,28 +5,28 @@
     <div class="jumbotron jumbotron-fluid">
       <div class="col-12">
         <div class="row d-flex align-items-center">
-          <div class="col-9 text-left" *ngIf="(nodeService.wallet(walletInfo) | async) as _wallet">
+          <div class="col-9 text-left" *ngIf="(nodeService.wallet(globalService.currentWallet) | async) as _wallet">
             <h5>Balance</h5>
             <p class="lead">
               <strong>{{ (_wallet.amountConfirmed | coinNotation) || (0 | coinNotation) }}</strong>
-              <small class="text-uppercase"> {{ coinUnit }}</small>
+              <small class="text-uppercase"> {{ globalService.coinUnit }}</small>
             </p>
             <ul class="list-unstyled">
-              <!-- <li><strong>{{ (awaitingMaturity | coinNotation) || (0 | coinNotation) }}</strong><em> (staking)</em></li> -->
-              <li><strong>{{ (_wallet.amountUnconfirmed | coinNotation) || (0 | coinNotation) }}</strong><em>
-                (unconfirmed)</em></li>
-              <!-- <li><strong>805.65900153</strong> <em>(staking)</em></li> -->
+              <li>
+                <strong>{{ (_wallet.amountUnconfirmed | coinNotation) || (0 | coinNotation) }}</strong>
+                <em> (unconfirmed)</em>
+              </li>
             </ul>
           </div>
           <!-- /col-->
           <div class="col text-right">
             <p>
-              <a class="btn btn-lg btn-primary" (click)="openReceiveDialog()" [class.btn-secondary]="sidechainEnabled">
+              <a class="btn btn-lg btn-primary" (click)="openReceiveDialog()" [class.btn-secondary]="globalService.getSidechainEnabled()">
                 <i class="icon lnr lnr-arrow-down-circle"></i>
                 <span class="text-center text-uppercase">Receive</span>
               </a>
               <a class="btn btn-lg btn-primary ml-2" (click)="openSendDialog()"
-                 [class.btn-secondary]="sidechainEnabled">
+                 [class.btn-secondary]="globalService.getSidechainEnabled()">
                 <i class="icon lnr lnr-arrow-up-circle"></i>
                 <span class="text-center text-uppercase">Send</span>
               </a>
@@ -40,65 +40,12 @@
     </div>
     <!-- /JUMBOTRON-->
 
-    <app-transactions [transactions]="transactionArray"></app-transactions>
+    <app-transactions></app-transactions>
 
-    <ng-container *ngIf="(nodeService.wallet(walletInfo) | async) as _wallet">
-      <section class="col-4" *ngIf="_wallet.hasBalance && !sidechainEnabled">
-        <div class="card" *ngIf="!stakingEnabled">
-          <div class="card-header">Staking</div>
-          <div class="card-body">
-            <div class="alert alert-info text-center" role="alert">Unlock your wallet to start staking</div>
-            <form class="form-inline mb-3" [formGroup]="stakingForm">
-              <div class="form-group">
-                <label for="inputPassword" class="sr-only">Password</label>
-                <input type="password" formControlName="walletPassword" class="form-control" id="inputPassword"
-                       placeholder="Password">
-                <button type="button" id="flipcard" class="btn btn-outline-primary ml-2"
-                        [disabled]="isStarting || !stakingForm.valid" (click)="startStaking()">Unlock
-                </button>
-              </div>
-            </form>
-          </div>
-        </div>
-        <div id="staking" class="card" *ngIf="stakingEnabled">
-          <div class="card-header">
-            <span class="d-inline-block col-10 row">Staking</span>
-            <!-- <div class="toggle d-inline-block" id="switch">
-              <div class='toggle-text-off'>OFF</div>
-              <div class='glow-comp'></div>
-              <div class='toggle-button'></div>
-              <div class='toggle-text-on'>ON</div>
-            </div> -->
-          </div>
-          <div class="card-body">
-            <ul class="list-unstyled mb-3">
-              <li *ngIf="isStarting">Waiting for staking to start...</li>
-              <div *ngIf="!isStarting">
-                <li>Staking weight:
-                  <strong>{{ stakingWeight | coinNotation | number: '1.0-0' }} {{ coinUnit }}</strong></li>
-                <li>Coins awaiting maturity:
-                  <strong>{{ awaitingMaturity | coinNotation | number: '1.0-0' }} {{ coinUnit }}</strong></li>
-                <li>Network weight:
-                  <strong>{{ netStakingWeight | coinNotation | number: '1.0-0' }} {{ coinUnit }}</strong></li>
-                <li class="mt-2">Expected reward time is:</li>
-                <li><strong>{{ dateTime }}</strong></li>
-              </div>
-            </ul>
-            <button type="button" class="btn btn-outline-light mb-3" [disabled]="isStopping"
-                    (click)="stopStaking()">
-              Stop Staking
-            </button>
-          </div>
-        </div>
-      </section>
-    </ng-container>
-    <!-- STAKING -->
+    <app-staking></app-staking>
 
-    </div>
     <status-bar></status-bar>
   </section>
-  <!-- /CONTENT -->
-</section>
-<!-- /CONTAINER-->
+  <!-- /CONTAINER-->
 </section>
 <!-- /GUI-->
