+++ StratisCore.UI/src/app/wallet/dashboard/dashboard.component.ts
@@ -1,20 +1,10 @@
-import { Component, OnInit, OnDestroy, Input } from '@angular/core';
-import { NgbModal, NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
-import { FormGroup, FormControl, Validators, FormBuilder } from '@angular/forms';
-
+import { Component, OnInit } from '@angular/core';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { ApiService } from '@shared/services/api.service';
 import { GlobalService } from '@shared/services/global.service';
-import { ModalService } from '@shared/services/modal.service';
-import { WalletInfo } from '@shared/models/wallet-info';
-import { TransactionInfo } from '@shared/models/transaction-info';
-
 import { SendComponent } from '../send/send.component';
 import { ReceiveComponent } from '../receive/receive.component';
-
-import { Subscription } from 'rxjs';
-import { Router } from '@angular/router';
 import { StratisNodeService } from "@shared/services/real-time/stratis-node.service";
-import { SecondsToStringPipe } from "@shared/pipes/seconds-to-string.pipe";
 
 @Component({
   selector: 'dashboard-component',
@@ -22,207 +12,22 @@ import { SecondsToStringPipe } from "@shared/pipes/seconds-to-string.pipe";
   styleUrls: ['./dashboard.component.css']
 })
 
-export class DashboardComponent implements OnInit, OnDestroy {
-  private stakingForm: FormGroup;
-  private walletHistorySubscription: Subscription;
-  private stakingInfoSubscription: Subscription;
-
+export class DashboardComponent implements OnInit {
   constructor(
-    private nodeService : StratisNodeService,
+    private nodeService: StratisNodeService,
     private apiService: ApiService,
     private globalService: GlobalService,
-    private modalService: NgbModal,
-    private genericModalService: ModalService,
-    private router: Router,
-    private fb: FormBuilder) {
-
-    this.buildStakingForm();
+    private modalService: NgbModal) {
   }
 
-  public sidechainEnabled: boolean;
-  public walletName: string;
-  public coinUnit: string;
-  public spendableBalance: number;
-  public transactionArray: TransactionInfo[];
-  public stakingEnabled: boolean;
-  public stakingActive: boolean;
-  public stakingWeight: number;
-  public awaitingMaturity: number = 0;
-  public netStakingWeight: number;
-  public expectedTime: number;
-  //public dateTime: string;
-  public isStarting: boolean;
-  public isStopping: boolean;
-
-  public walletInfo : WalletInfo;
-
-  ngOnInit() {
-    this.walletInfo = new WalletInfo(this.globalService.getWalletName());
-    this.sidechainEnabled = this.globalService.getSidechainEnabled();
-    this.walletName = this.globalService.getWalletName();
-    this.coinUnit = this.globalService.getCoinUnit();
-    this.startSubscriptions();
+  public ngOnInit() {
   };
 
-  ngOnDestroy() {
-    this.cancelSubscriptions();
-  }
-
-  private buildStakingForm(): void {
-    this.stakingForm = this.fb.group({
-      "walletPassword": ["", Validators.required]
-    });
-  }
-
-
   public openSendDialog() {
-    const modalRef = this.modalService.open(SendComponent, {backdrop: "static", keyboard: false});
+    this.modalService.open(SendComponent, {backdrop: "static", keyboard: false});
   }
 
   public openReceiveDialog() {
-    const modalRef = this.modalService.open(ReceiveComponent, {backdrop: "static", keyboard: false});
+    this.modalService.open(ReceiveComponent, {backdrop: "static", keyboard: false});
   };
-
-
-  // todo: add history in separate service to make it reusable
-  private getHistory() {
-    const walletInfo = new WalletInfo(this.globalService.getWalletName());
-    let historyResponse;
-    this.walletHistorySubscription = this.apiService.getWalletHistory(walletInfo)
-      .subscribe(
-        response => {
-          // TO DO - add account feature instead of using first entry in array
-          if (!!response.history && response.history[0].transactionsHistory.length > 0) {
-            historyResponse = response.history[0].transactionsHistory;
-            this.getTransactionInfo(historyResponse);
-          }
-        },
-        error => {
-          if (error.status === 0) {
-            this.cancelSubscriptions();
-          } else if (error.status >= 400) {
-            if (!error.error.errors[0].message) {
-              this.cancelSubscriptions();
-              this.startSubscriptions();
-            }
-          }
-        }
-      )
-    ;
-  };
-
-  private getTransactionInfo(transactions: any) {
-    this.transactionArray = [];
-
-    for (let transaction of transactions) {
-      let transactionType;
-      if (transaction.type === "send") {
-        transactionType = "sent";
-      } else if (transaction.type === "received") {
-        transactionType = "received";
-      } else if (transaction.type === "staked") {
-        transactionType = "staked";
-      }
-      let transactionId = transaction.id;
-      let transactionAmount = transaction.amount;
-      let transactionFee;
-      if (transaction.fee) {
-        transactionFee = transaction.fee;
-      } else {
-        transactionFee = 0;
-      }
-      let transactionConfirmedInBlock = transaction.confirmedInBlock;
-      let transactionTimestamp = transaction.timestamp;
-
-      this.transactionArray.push(new TransactionInfo(transactionType, transactionId, transactionAmount, transactionFee, transactionConfirmedInBlock, transactionTimestamp));
-    }
-  }
-
-  private startStaking() {
-    this.isStarting = true;
-    this.isStopping = false;
-    const walletData = {
-      name: this.globalService.getWalletName(),
-      password: this.stakingForm.get('walletPassword').value
-    };
-    this.apiService.startStaking(walletData)
-      .subscribe(
-        response => {
-          this.stakingEnabled = true;
-          this.stakingForm.patchValue({walletPassword: ""});
-          this.getStakingInfo();
-        },
-        error => {
-          this.isStarting = false;
-          this.stakingEnabled = false;
-          this.stakingForm.patchValue({walletPassword: ""});
-        }
-      )
-    ;
-  }
-
-  private stopStaking() {
-    this.isStopping = true;
-    this.isStarting = false;
-    this.apiService.stopStaking()
-      .subscribe(
-        response => {
-          this.stakingEnabled = false;
-        }
-      )
-    ;
-  }
-
-  private getStakingInfo() {
-    this.stakingInfoSubscription = this.apiService.getStakingInfo()
-      .subscribe(
-        response => {
-          const stakingResponse = response;
-          this.stakingEnabled = stakingResponse.enabled;
-          this.stakingActive = stakingResponse.staking;
-          this.stakingWeight = stakingResponse.weight;
-          this.netStakingWeight = stakingResponse.netStakeWeight;
-          this.awaitingMaturity = (this.unconfirmedBalance + this.confirmedBalance) - this.spendableBalance;
-          this.expectedTime = stakingResponse.expectedTime;
-          this.dateTime = new SecondsToStringPipe().transform(this.expectedTime);
-          if (this.stakingActive) {
-            this.isStarting = false;
-          } else {
-            this.isStopping = false;
-          }
-        }, error => {
-          if (error.status === 0) {
-            this.cancelSubscriptions();
-          } else if (error.status >= 400) {
-            if (!error.error.errors[0].message) {
-              this.cancelSubscriptions();
-              this.startSubscriptions();
-            }
-          }
-        }
-      )
-    ;
-  }
-
-  private cancelSubscriptions() {
-    // if (this.walletBalanceSubscription) {
-    //   this.walletBalanceSubscription.unsubscribe();
-    // }
-
-    if (this.walletHistorySubscription) {
-      this.walletHistorySubscription.unsubscribe();
-    }
-
-    if (this.stakingInfoSubscription) {
-      this.stakingInfoSubscription.unsubscribe();
-    }
-  }
-
-  private startSubscriptions() {
-    //this.getWalletBalance();
-    this.getHistory();
-    if (!this.sidechainEnabled) {
-      this.getStakingInfo();
-    }
-  }
 }
